Split native and custom controller code into separate files.
