+++ src/ts/instagram/videoPlayer.ts
@@ -3,7 +3,9 @@ import {VideoType} from "./videoType";
 import {PlaybackManager} from "./playbackManager";
 import {Settings} from "../shared/settings";
 import {VideoControlMode} from "../shared/videoControlMode";
-import {Browser} from "../shared/browser";
+import {VideoController} from "./controller/videoController";
+import {NativeVideoController} from "./controller/nativeVideoController";
+import {CustomVideoController} from "./controller/customVideoController";
 
 // The custom video player for Instagram video tags.
 export class VideoPlayer {
@@ -81,14 +83,14 @@ export class VideoPlayer {
 
     // Handles video play event.
     private onPlay = () => {
-        this.updatePlayControl();
+        this.videoController?.onPlay();
 
         this.playbackManager.notifyVideoPlay(this.videoElement);
     }
 
     // Handles video pause event.
     private onPause = () => {
-        this.updatePlayControl();
+        this.videoController?.onPause();
 
         // If not registered, the site will show a 'you need to log in to replay' banner.
         // This makes sure it is removed, so you regain control.
@@ -99,12 +101,12 @@ export class VideoPlayer {
 
     // Handles video time update event.
     private onTimeUpdate = () => {
-        this.updatePositionControl();
+        this.videoController?.onTimeUpdate();
     }
 
     // Handles video volume changes.
     private onVolumeChange = () => {
-        this.updateVolumeControl();
+        this.videoController?.onVolumeChange();
 
         // We don't want to react to volume changes from the page itself.
         //if (!event.isTrusted) return;
@@ -114,22 +116,22 @@ export class VideoPlayer {
 
     // Handles fullscreen changes.
     private onFullscreenChange = () => {
-        this.updateFullscreenControl();
+        this.videoController?.onFullscreenChange();
     }
 
     // Handles Picture-in-Picture changes.
     private onPictureInPictureChange = () => {
-        this.updatePictureInPictureControl();
+        this.videoController?.onPictureInPictureChange();
     }
 
     // Mouse enters the player element.
     private onMouseEnter = () => {
-        this.setHover(true);
+        this.videoController?.setHover(true);
     }
 
     // Mouse leaves the player element.
     private onMouseLeave = () => {
-        this.setHover(false);
+        this.videoController?.setHover(false);
     }
 
     //#endregion
@@ -143,22 +145,22 @@ export class VideoPlayer {
     public isEmbedded: boolean = false;
 
     // The native root element of the video area with all controls.
-    private videoRootElement: HTMLElement | undefined;
+    public videoRootElement: HTMLElement | undefined;
 
     // The native Instagram overlay element.
-    private overlayElement: HTMLElement | undefined;
+    public overlayElement: HTMLElement | undefined;
 
     // The native Instagram reply element for stories.
-    private replyElement: HTMLElement | undefined;
+    public replyElement: HTMLElement | undefined;
 
     // The native clickable areas for mobile Story controls.
-    private clickEventElement: HTMLElement | undefined;
+    public clickEventElement: HTMLElement | undefined;
 
     // The native mute button.
-    private muteElement: HTMLElement | undefined;
+    public muteElement: HTMLElement | undefined;
 
     // The overlay elements used for Reels on mobile.
-    private mobileOverlayElement: HTMLElement | undefined;
+    public mobileOverlayElement: HTMLElement | undefined;
 
 
     // Detects the video type and finds all native components.
@@ -273,319 +275,29 @@ export class VideoPlayer {
 
     // #region Controls
 
-    // Created elements by the video controls.
-    private videoControlElement: HTMLElement | undefined;
-    private playButtonElement: HTMLButtonElement | undefined;
-    private seekBarProgressElement: HTMLElement | undefined;
-    private positionTextElement: HTMLElement | undefined;
-    private muteButtonElement: HTMLButtonElement | undefined;
-    private volumeBarProgressElement: HTMLElement | undefined;
-    private fullscreenButtonElement: HTMLButtonElement | undefined;
-    private pictureInPictureButtonElement: HTMLButtonElement | undefined;
-
-    // Cache of the created image tags for the icons.
-    private static imagePlay = Browser.getUrl('images/play.svg');
-    private static imagePause = Browser.getUrl('images/pause.svg');
-    private static imageFullscreenEnter = Browser.getUrl('images/fullscreen-enter.svg');
-    private static imageFullscreenExit = Browser.getUrl('images/fullscreen-exit.svg');
-    private static imageSpeakerOn = Browser.getUrl('images/speaker-on.svg');
-    private static imageSpeakerOff = Browser.getUrl('images/speaker-off.svg');
-    private static imagePictureInPictureEnter = Browser.getUrl('images/picture-in-picture-enter.svg');
-    private static imagePictureInPictureExit = Browser.getUrl('images/picture-in-picture-exit.svg');
-
-    // Changes the icon of a button containing an image element.
-    private static setButtonIcon(button: HTMLButtonElement, url: string) {
-        const img = button.firstChild as HTMLImageElement;
-        if (!img) return;
-        img.src = url;
-    }
+    // The current created controller.
+    private videoController: VideoController | undefined;
 
     // Creates the video controls.
     private createVideoControl() {
         switch (Settings.shared.videoControlMode) {
             case VideoControlMode.native:
-                this.createNativeVideoControl();
+                this.videoController = new NativeVideoController(this);
                 break;
             case VideoControlMode.custom:
-                this.createCustomVideoControl();
+                this.videoController = new CustomVideoController(this);
                 break;
         }
-    }
-
-    // Adjust the control bar height for the background and all native elements.
-    private adjustVideoControlHeight(controlHeight: number) {
-        if (!this.overlayElement) return;
-
-        // Removes the height of the controls from the inner overlay to not block mouse clicks.
-        if (this.overlayElement.firstChild instanceof HTMLElement) {
-            this.overlayElement.firstChild.style.height = `calc(100% - ${controlHeight}px)`;
-        }
-
-        // For Stories, we also add a margin to the reply element to not overlay the controls.
-        if (this.replyElement) {
-            // The social controls in mobile Stories are placed below the post and don't overlap by default.
-            if (this.videoType !== VideoType.mobileStory) {
-                this.replyElement.style.marginBottom = `${controlHeight}px`;
-            }
-        }
-
-        // For Reels on mobile, these are different elements with absolut position.
-        if (this.mobileOverlayElement) {
-            this.mobileOverlayElement.style.bottom = `${controlHeight}px`;
-        }
-
-        // If clickable overlays are used, we want to add a margin, so we can interact with our controls without
-        // activating the overlay buttons.
-        if (this.clickEventElement) {
-            this.clickEventElement.style.marginBottom = `${controlHeight}px`;
-        }
-
-        // Hide the native mute button.
-        if (this.muteElement) {
-            this.muteElement.style.display = 'none';
-        }
-
-        // Adjusting the controller background.
-        if (this.videoControlElement) {
-            this.videoControlElement.style.height = `${controlHeight}px`;
-        }
-    }
-
-    // Create the control background.
-    private createVideoControlBackground() {
-        if (!this.overlayElement) return;
-
-        this.videoControlElement = document.createElement("div");
-        this.videoControlElement.classList.add("ivc-controls");
-        if (this.videoType === VideoType.reel) {
-            this.videoControlElement.classList.add("ivc-reel");
-        }
-        if (this.videoType === VideoType.story) {
-            this.videoControlElement.classList.add("ivc-story");
-        }
-        this.overlayElement.appendChild(this.videoControlElement);
-    }
-
-    // Enables the native browser controls.
-    private createNativeVideoControl() {
-        if (!this.overlayElement) return;
-
-        // The controls in Chrome are higher than in Firefox.
-        const controlHeight = window.chrome ? 70 : 40;
-
-        this.createVideoControlBackground();
-        this.adjustVideoControlHeight(controlHeight);
-        this.updateControlBarVisibility();
-    }
-
-    // Create the video controls with play/pause buttons, seekbar and volume control.
-    private createCustomVideoControl() {
-        if (!this.overlayElement) return;
-        const video = this.videoElement;
-
-        const controlHeight = 32;
-
-        this.createVideoControlBackground();
-        this.adjustVideoControlHeight(controlHeight);
-
-        // Creating the actual player...
-        if (!this.videoControlElement) return;
-
-        const contentElement = document.createElement("div");
-        contentElement.classList.add("ivc-controls-content");
-        this.videoControlElement.appendChild(contentElement);
-
-        // Play button
-        this.playButtonElement = document.createElement("button");
-        this.playButtonElement.classList.add("ivc-control-element", "ivc-icon-button");
-        this.playButtonElement.appendChild(document.createElement("img"));
-        contentElement.appendChild(this.playButtonElement);
-
-        this.playButtonElement.onclick = () => {
-            if (video.paused) {
-                video.play().then();
-            } else {
-                video.pause();
-            }
-        };
-
-        // Position text
-        this.positionTextElement = document.createElement("div");
-        this.positionTextElement.classList.add("ivc-control-element", "ivc-control-text");
-        contentElement.appendChild(this.positionTextElement);
-
-        // Seekbar
-        const elementSeekbar = document.createElement("div");
-        elementSeekbar.classList.add("ivc-control-element", "ivc-control-bar", "ivc-seek-bar");
-        contentElement.appendChild(elementSeekbar);
-
-        const elementSeekbarBackground = document.createElement("div");
-        elementSeekbarBackground.classList.add("ivc-control-bar-background");
-        elementSeekbar.appendChild(elementSeekbarBackground);
-
-        this.seekBarProgressElement = document.createElement("div");
-        this.seekBarProgressElement.classList.add("ivc-control-bar-progress");
-        elementSeekbarBackground.appendChild(this.seekBarProgressElement);
-
-        VideoPlayer.addDragEventToBar(elementSeekbar, elementSeekbarBackground, this.seekBarProgressElement,
-            /* invokeOnDrag */ false, (value) => {
-                video.currentTime = value * video.duration;
-            });
-
-        // Mute
-        this.muteButtonElement = document.createElement("button");
-        this.muteButtonElement.classList.add("ivc-control-element", "ivc-icon-button");
-        this.muteButtonElement.appendChild(document.createElement("img"));
-        contentElement.appendChild(this.muteButtonElement);
-
-        this.muteButtonElement.onclick = () => {
-            video.muted = !video.muted;
-        };
-
-        // Volume
-        const elementVolume = document.createElement("div");
-        elementVolume.classList.add("ivc-control-element", "ivc-control-bar", "ivc-volume-bar");
-        contentElement.appendChild(elementVolume);
-
-        const elementVolumeBackground = document.createElement("div");
-        elementVolumeBackground.classList.add("ivc-control-bar-background");
-        elementVolume.appendChild(elementVolumeBackground);
-
-        this.volumeBarProgressElement = document.createElement("div");
-        this.volumeBarProgressElement.classList.add("ivc-control-bar-progress");
-        elementVolumeBackground.appendChild(this.volumeBarProgressElement);
-
-        VideoPlayer.addDragEventToBar(elementVolume, elementVolumeBackground, this.volumeBarProgressElement,
-            /* invokeOnDrag */ true, (value) => {
-                video.volume = value;
-                video.muted = video.volume <= 0;
-            });
-
-        // Picture-in-Picture
-        this.pictureInPictureButtonElement = document.createElement("button");
-        this.pictureInPictureButtonElement.classList.add("ivc-control-element", "ivc-icon-button");
-        this.pictureInPictureButtonElement.appendChild(document.createElement("img"));
-        contentElement.appendChild(this.pictureInPictureButtonElement);
-
-        this.pictureInPictureButtonElement.onclick = () => {
-            if (!this.videoElement) return;
-
-            // Toggle Picture-in-picture
-            if (document.pictureInPictureElement) {
-                document.exitPictureInPicture().then();
-            } else {
-                this.videoElement.requestPictureInPicture().then();
-            }
-        };
-
-        // Full screen
-        this.fullscreenButtonElement = document.createElement("button");
-        this.fullscreenButtonElement.classList.add("ivc-control-element", "ivc-icon-button");
-        this.fullscreenButtonElement.appendChild(document.createElement("img"));
-        contentElement.appendChild(this.fullscreenButtonElement);
-
-        this.fullscreenButtonElement.onclick = () => {
-            if (!this.videoRootElement) return;
-
-            // Toggle fullscreen
-            if (document.fullscreenElement) {
-                document.exitFullscreen().then();
-            } else {
-                this.videoRootElement.requestFullscreen().then();
-            }
-        };
-
-        // Init update
-        this.updatePlayControl();
-        this.updatePositionControl();
-        this.updateVolumeControl();
-        this.updateFullscreenControl();
-        this.updatePictureInPictureControl();
-        this.updateControlBarVisibility();
+        this.videoController?.create();
     }
 
     // Removes the video controls
     private removeVideoControl() {
-        const video = this.videoElement;
-
-        // Remove controls
-        video.controls = false;
-
-        // Restore overlay margin
-        if (this.overlayElement && this.overlayElement.firstChild instanceof HTMLElement) {
-            this.overlayElement.firstChild.style.height = '';
-        }
-
-        // Restore reply controls
-        if (this.replyElement) {
-            this.replyElement.style.marginBottom = '';
-        }
-
-        // Restore Reels controls
-        if (this.mobileOverlayElement) {
-            this.mobileOverlayElement.style.bottom = '';
-        }
-
-        // Restore click handler
-        if (this.clickEventElement) {
-            this.clickEventElement.style.marginBottom = '';
-        }
-
-        // Restore original mute button
-        if (this.muteElement) {
-            this.muteElement.style.display = '';
-        }
-
-        if (this.videoControlElement) {
-            this.videoControlElement.remove();
-            this.videoControlElement = undefined;
-        }
+        if (!this.videoController) return;
+        this.videoController.remove();
+        this.videoController = undefined;
     }
 
-    private updatePlayControl() {
-        if (!this.playButtonElement) return;
-        VideoPlayer.setButtonIcon(this.playButtonElement, this.videoElement.paused ?
-            VideoPlayer.imagePlay : VideoPlayer.imagePause);
-    }
-
-    private updatePositionControl() {
-        if (!this.seekBarProgressElement || !this.positionTextElement) return;
-
-        const progress = this.videoElement.currentTime / this.videoElement.duration;
-        this.seekBarProgressElement.style.width = `${Math.round(progress * 100)}%`
-
-        this.positionTextElement.innerText =
-            `${Utils.formatTime(this.videoElement.currentTime)} / ${Utils.formatTime(this.videoElement.duration)}`;
-
-        this.setElementVisibility(this.positionTextElement, Settings.shared.showTimeCodeText);
-    }
-
-    private updateVolumeControl() {
-        if (!this.muteButtonElement || !this.volumeBarProgressElement) return;
-        VideoPlayer.setButtonIcon(this.muteButtonElement, this.videoElement.muted ?
-            VideoPlayer.imageSpeakerOff : VideoPlayer.imageSpeakerOn);
-        this.volumeBarProgressElement.style.width = `${Math.round(this.videoElement.volume * 100)}%`
-    }
-
-    private updateFullscreenControl() {
-        if (!this.fullscreenButtonElement) return;
-        VideoPlayer.setButtonIcon(this.fullscreenButtonElement, document.fullscreenElement ?
-            VideoPlayer.imageFullscreenExit : VideoPlayer.imageFullscreenEnter);
-
-        // Only show the fullscreen button if it is available in the current context. It can be disabled by iframes.
-        this.setElementVisibility(this.fullscreenButtonElement,
-            document.fullscreenEnabled && Settings.shared.showFullscreenButton);
-    }
-
-    private updatePictureInPictureControl() {
-        if (!this.pictureInPictureButtonElement) return;
-        VideoPlayer.setButtonIcon(this.pictureInPictureButtonElement, document.pictureInPictureElement ?
-            VideoPlayer.imagePictureInPictureExit : VideoPlayer.imagePictureInPictureEnter);
-
-        // Only show the PiP button if it is available in the current context. It is not available in Firefox!
-        this.setElementVisibility(this.pictureInPictureButtonElement,
-            document.pictureInPictureEnabled && Settings.shared.showPictureInPictureButton);
-    }
 
     private removeNotRegisteredOverlayElement() {
         const nativeControlsElement = this.overlayElement?.firstChild as HTMLElement;
@@ -599,47 +311,17 @@ export class VideoPlayer {
         }
     }
 
-    // Mouse is hovering the player element.
-    private hover: boolean = false;
-
-    private setHover(hover: boolean) {
-        this.hover = hover;
-        this.updateControlBarVisibility();
-    }
-
     // Checks if the mouse already hovers over the video on initialization. Otherwise, the controls would only show up
     // when the mouse exits and enters again.
     private initHover() {
         if (!this.videoRootElement) return;
         const hover = this.videoRootElement.matches(':hover');
-        this.setHover(hover);
-    }
-
-    // Changes the visibility of the video controls.
-    private setControlBarVisibility(visibility: boolean) {
-        switch (Settings.shared.videoControlMode) {
-            case VideoControlMode.native:
-                if (!this.videoElement) return;
-                this.videoElement.controls = visibility;
-                break;
-            case VideoControlMode.custom:
-                if (!this.videoControlElement) return;
-                this.videoControlElement.classList.toggle('hidden', !visibility);
-                break;
-        }
-    }
-
-    private updateControlBarVisibility() {
-        const visibility = !Settings.shared.autoHideControlBar || this.hover;
-        this.setControlBarVisibility(visibility);
+        this.videoController?.setHover(hover);
     }
 
     // Is called when any control setting was changed. We should update all dynamic controls.
     public updateControlSetting() {
-        this.updatePositionControl();
-        this.updateFullscreenControl();
-        this.updatePictureInPictureControl();
-        this.updateControlBarVisibility();
+        this.videoController?.onUpdateSettings();
     }
 
     // Is called when the control mode was changed. This rebuilds the UI.
@@ -648,61 +330,5 @@ export class VideoPlayer {
         this.attach();
     }
 
-    // Changes the visibility of a control element.
-    private setElementVisibility(element: HTMLElement, visible: boolean) {
-        element.style.display = visible ? 'block' : 'none';
-    }
-
-    // Handles click and drag events to the bars elements (e.g. seekbar, volume bar).
-    private static addDragEventToBar(element: HTMLElement, elementBackground: HTMLElement, elementProgress: HTMLElement,
-                                     invokeOnDrag: boolean, callback: (value: number) => void) {
-
-        // Sub function to submit the current bar value
-        const onValueChanged = (event: MouseEvent | TouchEvent, invoke: boolean) => {
-            const rect = elementBackground.getBoundingClientRect();
-
-            const clientX = event instanceof MouseEvent ? event.clientX : event.changedTouches[0].clientX;
-            const relativeX = clientX - rect.left;
-            const value = Math.max(Math.min(relativeX / rect.width, 1), 0);
-
-            elementProgress.style.width = `${Math.round(value * 100)}%`;
-
-            if (invoke)
-                callback(value);
-        };
-
-        // Handle click event
-
-        element.addEventListener("click", (event) => {
-            onValueChanged(event, true);
-        });
-
-        // Handle drag event
-
-        let isDragging = false;
-        const onDragStart = () => {
-            isDragging = true;
-        }
-        const onDragEnd = (event: MouseEvent | TouchEvent) => {
-            if (!isDragging) return;
-            onValueChanged(event, true);
-            isDragging = false;
-        }
-        const onDrag = (event: MouseEvent | TouchEvent) => {
-            if (!isDragging) return;
-            onValueChanged(event, invokeOnDrag);
-        }
-
-        element.addEventListener("mousedown", onDragStart);
-        element.addEventListener("touchstart", onDragStart);
-
-        element.addEventListener("mousemove", onDrag);
-        element.addEventListener("touchmove", onDrag);
-
-        element.addEventListener("mouseup", onDragEnd);
-        element.addEventListener("touchend", onDragEnd);
-        element.addEventListener("mouseleave", onDragEnd);
-    }
-
-    //#endregion
+    //#endregion Controls
 }
